Configure bundler with rails lts credentials
Because other docker images built off of this base image set BUNDLE_PATH
to a volume, we cannot install railslts here and assume it will be used
when installing the bundle.

I think it is better to configure bundler with the credentials anyway.
In the case that railslts releases a new version, we can simply bundle
install vs rebuilding the base image.